+++ src/core/component/directives/attrs/const.ts
@@ -13,32 +13,6 @@ export const
 export const
 	handlers = new WeakMap<object, Map<unknown, Function>>();
 
-export const keyModifiers = Object.createDict({
-	enter: true,
-	tab: true,
-	delete: true,
-	esc: true,
-	space: true,
-	up: true,
-	down: true,
-	left: true,
-	right: true
-});
-
-export const modifiers = Object.createDict({
-	left: true,
-	middle: true,
-	right: true,
-	ctrl: true,
-	alt: true,
-	shift: true,
-	meta: true,
-	exact: true,
-	self: true,
-	prevent: true,
-	stop: true
-});
-
 export const classAttrs = Object.createDict({
 	class: 'class',
 	'^class': 'class',