+++ src/core/component/directives/attrs/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.??? (2024-??-??)
+
+#### :house: Internal
+
+* Moved the `parseEventListener` function to common directive helpers
+
 ## v4.0.0-beta.111 (2024-07-18)
 
 #### :bug: Bug Fix